+++ frontend/src/contexts/SimulationContext.tsx
@@ -1,13 +1,13 @@
-import React, { createContext, useState } from 'react';
+import React, { createContext, useState, useContext } from 'react';
 import type { RunResponse } from '@utils/api';
 
-export interface SimulationContextType {
+interface SimulationContextType {
   simulationResult: RunResponse | null;
   setSimulationResult: (result: RunResponse | null) => void;
   isSimulationReady: boolean;
 }
 
-export const SimulationContext = createContext<SimulationContextType | undefined>(undefined);
+const SimulationContext = createContext<SimulationContextType | undefined>(undefined);
 
 export const SimulationProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [simulationResult, setSimulationResult] = useState<RunResponse | null>(null);
@@ -23,4 +23,13 @@ export const SimulationProvider: React.FC<{ children: React.ReactNode }> = ({ ch
       {children}
     </SimulationContext.Provider>
   );
+};
+
+// eslint-disable-next-line react-refresh/only-export-components
+export const useSimulation = (): SimulationContextType => {
+  const context = useContext(SimulationContext);
+  if (context === undefined) {
+    throw new Error('useSimulation must be used within a SimulationProvider');
+  }
+  return context;
 };
